Create dirs and refactor comp into them
